+++ src/types.ts
@@ -3,3 +3,5 @@ import type { Document } from 'mongodb'
 export interface SyncOptions {
   mapper?: (doc: Document) => Document
 }
+
+export type Events = 'process' | 'error'